before:
```
benchmarking normalization of benchmark/tests/PipelinesViaFolds.hs
time                 1.066 s    (1.052 s .. 1.093 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.065 s    (1.057 s .. 1.070 s)
std dev              8.007 ms   (2.932 ms .. 10.99 ms)
variance introduced by outliers: 19% (moderately inflated)
```

after:
```
benchmarking normalization of benchmark/tests/PipelinesViaFolds.hs
time                 966.0 ms   (923.8 ms .. 986.9 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 973.8 ms   (967.6 ms .. 977.4 ms)
std dev              6.027 ms   (1.990 ms .. 8.113 ms)
variance introduced by outliers: 19% (moderately inflated)
```